requirements.txt
absl-py==1.3.0
appnope==0.1.4
APScheduler==3.10.0
asgiref==3.4.1
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.21.4
bleach==6.1.0
blinker==1.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
confluent-kafka==2.0.2
contourpy==1.0.5
cryptography==35.0.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.9
docopt==0.6.2
executing==2.0.1
Faker==17.0.0
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.37.4
greenlet==3.1.0
idna==3.3
imutils==0.5.4
iniconfig==1.1.1
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kafka-python==2.0.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.7
mediapipe-rpi4==0.8.8
mistune==3.0.2
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
paho-mqtt==1.6.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.5.0
platformdirs==4.2.1
pluggy==1.0.0
prompt-toolkit==3.0.43
protobuf==4.21.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pygame==2.0.3
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==26.0.2
referencing==0.35.0
requests==2.26.0
rpds-py==0.18.0
scipy==1.7.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.4.2
stack-data==0.6.3
tinycss2==1.3.0
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.7
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.1
Werkzeug==3.0.3
WTForms==3.1.2
yarg==0.1.9
